requirements.txt
python-multipart
pydantic
pydantic-settings
fastapi
asyncpg
uvicorn
SQLAlchemy
pandas
pdfplumber
aiofiles
pgvector
tabulate
google-genai
openai
